requirements.txt
flask
flask_sqlalchemy
flask_cors
gunicorn
psycopg2
cloudinary
python-dotenv
